weekly_milestones/week_06/etl_job/requirements.txt
numpy==1.24.3
pandas==2.0.1
pymongo==4.3.3
SQLAlchemy==2.0.12
psycopg2-binary==2.9.6
vaderSentiment==3.3.2
